PrestoBackEnd/requirements.txt
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-filters==0.11.1
mysqlclient==1.4.2.post1
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
